LayoutTests/platform/glib/accessibility/form-control-value-settable-expected.txt



This tests whether AXValue is writable for various form controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


text1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
text2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
number1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
number2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
textarea1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
textarea2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
slider1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
slider2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
checkbox1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
checkbox2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
radio1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
radio2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
submit1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is false
submit2
PASS document.activeElement == element is true
FAIL axElement.isAttributeSettable('AXValue') should be true. Was false.
combobox1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
combobox2
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
listbox1
PASS document.activeElement == element is true
PASS axElement.isAttributeSettable('AXValue') is true
listbox2
PASS document.activeElement == element is true
FAIL axElement.isAttributeSettable('AXValue') should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
